pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "dydactic"
version = "0.2.0"
description = "Validate iterables using Pydantic"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.9.2",
    "python-dateutil"
]
license = {text = "MIT License"}
keywords = ["iterable", "validation", "iterator", "model"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages]
find = {}

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=dydactic",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]